chore(inputs.powerdns): Don't run  flaky test on Windows
